LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-sticky-block-expected.txt

Subject

FAIL View timeline with sticky target, block axis. assert_equals: Effect at the midpoint of the active range: cover 0% to cover 100% expected "0.5" but got "0.7"
